+++ fugitive-darkness-api/src/main/java/com/zer0s2m/fugitivedarkness/api/handlers/ControllerApiGitJobGet.java
@@ -1,0 +1,65 @@
+package com.zer0s2m.fugitivedarkness.api.handlers;
+
+import com.zer0s2m.fugitivedarkness.models.GitJobModel;
+import com.zer0s2m.fugitivedarkness.repository.GitJobRepository;
+import com.zer0s2m.fugitivedarkness.repository.impl.GitJobRepositoryImpl;
+import io.netty.handler.codec.http.HttpResponseStatus;
+import io.vertx.core.Handler;
+import io.vertx.core.http.HttpHeaders;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.RoutingContext;
+import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.List;
+
+/**
+ * A request handler for getting all scheduled work for git repositories.
+ */
+final public class ControllerApiGitJobGet implements Handler<RoutingContext> {
+
+    static private final Logger logger = LoggerFactory.getLogger(ControllerApiGitJobGet.class);
+
+    /**
+     * Something has happened, so handle it.
+     *
+     * @param event the event to handle
+     */
+    @Override
+    public void handle(@NotNull RoutingContext event) {
+        final GitJobRepository gitJobRepository = new GitJobRepositoryImpl(event.vertx());
+
+        gitJobRepository
+                .findAll()
+                .onSuccess(ar -> {
+                    final JsonObject object = new JsonObject();
+                    final List<GitJobModel> gitJobs = gitJobRepository.mapTo(ar);
+
+                    object.put("success", true);
+                    object.put("gitJobs", gitJobs);
+
+                    gitJobRepository.closeClient();
+
+                    event
+                            .response()
+                            .putHeader(HttpHeaders.CONTENT_LENGTH, String.valueOf(object.toString().length()))
+                            .putHeader(HttpHeaders.CONTENT_TYPE, "application/json")
+                            .setStatusCode(HttpResponseStatus.OK.code())
+                            .write(object.toString());
+
+                    event.next();
+                })
+                .onFailure(error -> {
+                    gitJobRepository.closeClient();
+
+                    logger.error("Failure (DB): " + error.fillInStackTrace());
+
+                    event
+                            .response()
+                            .setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code())
+                            .end();
+                });
+    }
+
+}