+++ fugitive-darkness-models/src/main/resources/db/migration/V1_9_0__add_column_next_run_at_in_git_jobs.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "git_jobs"
+    ADD COLUMN next_run_at TIMESTAMP NULL;