+++ fugitive-darkness-plugin/fugitive-darkness-plugin-job/src/main/java/com/zer0s2m/fugitivedarkness/plugin/job/GitJobExecutor.java
@@ -1,0 +1,22 @@
+package com.zer0s2m.fugitivedarkness.plugin.job;
+
+/**
+ * The basic executor for performing tasks related to git repositories.
+ */
+public interface GitJobExecutor extends Runnable {
+
+    /**
+     * Set the project group.
+     *
+     * @param group The project group.
+     */
+    void setGroup(String group);
+
+    /**
+     * Set the name of the project.
+     *
+     * @param project The name of the project.
+     */
+    void setProject(String project);
+
+}