+++ fugitive-darkness-plugin/fugitive-darkness-plugin-job/src/main/java/com/zer0s2m/fugitivedarkness/plugin/job/GitJobExecutorOnetimeUse.java
@@ -1,0 +1,7 @@
+package com.zer0s2m.fugitivedarkness.plugin.job;
+
+/**
+ * An interface for implementing a scheduled task with the type {@link GitTypeJob#ONETIME_USE}.
+ */
+public interface GitJobExecutorOnetimeUse extends GitJobExecutor {
+}